+++ uristat/uristat-web/src/main/java/com/navercorp/pinpoint/uristat/web/controller/UriStatController.java
@@ -19,10 +19,10 @@ package com.navercorp.pinpoint.uristat.web.controller;
 import com.navercorp.pinpoint.common.server.util.time.ForwardRangeValidator;
 import com.navercorp.pinpoint.common.server.util.time.Range;
 import com.navercorp.pinpoint.common.server.util.time.RangeValidator;
+import com.navercorp.pinpoint.common.server.util.timewindow.TimePrecision;
 import com.navercorp.pinpoint.common.server.util.timewindow.TimeWindow;
 import com.navercorp.pinpoint.common.server.util.timewindow.TimeWindowSampler;
 import com.navercorp.pinpoint.common.server.util.timewindow.TimeWindowSlotCentricSampler;
-import com.navercorp.pinpoint.common.server.util.timewindow.TimePrecision;
 import com.navercorp.pinpoint.pinot.tenant.TenantProvider;
 import com.navercorp.pinpoint.uristat.web.chart.UriStatChartType;
 import com.navercorp.pinpoint.uristat.web.chart.UriStatChartTypeFactory;
@@ -100,8 +100,8 @@ public class UriStatController {
                 .setApplicationName(applicationName)
                 .setAgentId(agentId)
                 .setRange(range)
-                .setTimeSize((int) timeWindow.getWindowSlotSize())
-                .setTimePrecision(TimePrecision.newTimePrecision(TimeUnit.MILLISECONDS, (int) timeWindow.getWindowSlotSize()))
+                .setTimeSize(timeWindow.getWindowSlotSize())
+                .setTimePrecision(TimePrecision.newTimePrecision(TimeUnit.MILLISECONDS, timeWindow.getWindowSlotSize()))
                 .setOrderby(column)
                 .setDesc(isDesc)
                 .setLimit(count)
@@ -139,8 +139,8 @@ public class UriStatController {
                 .setAgentId(agentId)
                 .setUri(uri)
                 .setRange(timeWindow.getWindowRange())
-                .setTimeSize((int) timeWindow.getWindowSlotSize())
-                .setTimePrecision(TimePrecision.newTimePrecision(TimeUnit.MILLISECONDS, (int) timeWindow.getWindowSlotSize()))
+                .setTimeSize(timeWindow.getWindowSlotSize())
+                .setTimePrecision(TimePrecision.newTimePrecision(TimeUnit.MILLISECONDS, timeWindow.getWindowSlotSize()))
                 .build();
 
         UriStatChartType chartType = chartTypeFactory.valueOf(type.toLowerCase());
